syncpoint: make WaitForOneSeparateTimeoutTest less brittle
The test failed on FreeBSD because the thread wouldn't have started yet
or the timeout wouldn't have been triggered yet. As this test doesn't
test the timing but only makes sure that the other threads are not
affected by the thread that calls wait_for_one, increase the timeouts
and wait times to make the test more stable.